Index synchronization refactoring
